requirements.txt

openai>=1.50.0
requests>=2.31.0
pydantic>=2.8.2
python-dotenv>=1.0.1
APScheduler>=3.10.4
markdown>=3.6
Jinja2>=3.1.4
tweepy>=4.14.0
praw>=7.7.1
PyGithub>=2.4.0
tenacity>=9.0.0
PyJWT>=2.9.0
